+++ tests/test_arxiv.py
@@ -22,6 +22,7 @@ def test_find_arxiv_id():
         ('/URI(http://arxiv.org/abs/1011.2840)>>', '1011.2840'),
         ('/URI(http://arxiv.org/abs/1110.3658)>>', '1110.3658'),
         ('http://arxiv.org/abs/1110.3658>', '1110.3658'),
+        ('https://arxiv.org/abs/1110.3658>', '1110.3658'),
         ('http://arxiv.com/abs/1110.3658>', '1110.3658'),
         ('http://arxiv.org/1110.3658>', '1110.3658'),
     ]