+++ tools/ci-run-tests.sh
@@ -1,0 +1,5 @@
+#!/usr/bin/env bash
+
+python -m pytest papis/ tests/ --cov=papis
+mypy papis
+flake8 papis